RandomGenerator: Rykket alle random genereringer over i den rigtige klasse.
